@@ -26,7 +26,7 @@
     "bcrypt": "^5.0.1",
     "compression": "^1.7.3",
     "cookie": "^0.5.0",
-    "core-js": "^3.23.3",
+    "core-js": "^3.23.4",
     "etag": "^1.8.1",
     "express": "^4.17.2",
     "express-session": "^1.17.2",
@@ -38,7 +38,7 @@
     "morgan": "^1.10.0",
     "multer": "^1.4.4",
     "mysql2": "^2.3.3",
-    "nodemailer": "^6.7.2",
+    "nodemailer": "^6.7.7",
     "passport": "^0.6.0",
     "passport-discord": "^0.1.4",
     "passport-facebook": "^3.0.0",
@@ -47,40 +47,40 @@
     "passport-reddit": "^0.2.4",
     "passport-vkontakte": "^0.5.0",
     "ppfun-captcha": "^1.6.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.3.1",
-    "react-redux": "^7.2.6",
+    "react-redux": "^8.0.2",
     "react-responsive": "^8.2.0",
-    "react-stay-scrolled": "^7.4.0",
-    "react-toggle": "^4.1.2",
-    "redis": "^4.1.0",
+    "react-stay-scrolled": "^8.0.0",
+    "react-toggle": "https://github.com/aaronshaf/react-toggle",
+    "redis": "^4.2.0",
     "redux": "^4.1.2",
     "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
     "redux-thunk": "^2.4.1",
-    "sequelize": "^6.21.1",
+    "sequelize": "^6.21.3",
     "sharp": "^0.30.7",
     "startaudiocontext": "^1.2.1",
-    "three": "^0.141.0",
+    "three": "^0.142.0",
     "three-trackballcontrols": "^0.9.0",
     "ttag": "^1.7.24",
     "url-search-params-polyfill": "^8.1.1",
-    "winston": "^3.8.0",
+    "winston": "^3.8.1",
     "winston-daily-rotate-file": "^4.5.5",
     "ws": "^8.4.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.16.7",
-    "@babel/core": "^7.16.7",
+    "@babel/cli": "^7.18.6",
+    "@babel/core": "^7.18.6",
     "@babel/eslint-parser": "^7.16.5",
-    "@babel/node": "^7.16.7",
-    "@babel/plugin-proposal-throw-expressions": "^7.16.7",
-    "@babel/plugin-transform-flow-strip-types": "^7.16.7",
-    "@babel/plugin-transform-react-constant-elements": "^7.16.7",
-    "@babel/plugin-transform-react-inline-elements": "^7.16.7",
-    "@babel/preset-env": "^7.16.7",
-    "@babel/preset-react": "^7.16.7",
+    "@babel/node": "^7.18.6",
+    "@babel/plugin-proposal-throw-expressions": "^7.18.6",
+    "@babel/plugin-transform-flow-strip-types": "^7.18.6",
+    "@babel/plugin-transform-react-constant-elements": "^7.18.6",
+    "@babel/plugin-transform-react-inline-elements": "^7.18.6",
+    "@babel/preset-env": "^7.18.6",
+    "@babel/preset-react": "^7.18.6",
     "assets-webpack-plugin": "^7.1.1",
     "babel-loader": "^8.2.3",
     "babel-plugin-transform-react-pure-class-to-function": "^1.0.1",
@@ -88,14 +88,14 @@
     "babel-plugin-ttag": "^1.7.30",
     "clean-css": "^5.2.2",
     "clean-css-loader": "^4.1.1",
-    "copy-webpack-plugin": "^10.2.1",
+    "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.5.1",
-    "eslint": "^8.7.0",
+    "eslint": "^8.19.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-jsx-a11y": "^6.6.0",
-    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react": "^7.30.1",
     "generate-package-json-webpack-plugin": "^2.6.0",
     "mkdirp": "^1.0.4",
     "ttag-cli": "^1.9.3",
